intellicredit/requirements.txt
# ── API Layer ──
fastapi>=0.135.0
uvicorn[standard]>=0.41.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.5.0
pydantic>=2.12.0
pydantic-settings>=2.13.0

# ── Task Queue ──
# Keep redis client managed by `redis[hiredis]` below to avoid Celery extra pinning redis<6.
celery>=5.3.6
flower>=2.0.1

# ── Database Drivers ──
asyncpg>=0.31.0
redis[hiredis]>=7.2.0
neo4j>=5.17.0
elasticsearch[async]>=8.12.1
chromadb>=0.5.0

# ── Orchestration & Agents ──
langgraph>=1.0.0
langchain>=1.2.0
langchain-anthropic>=1.3.0
langchain-community>=0.4.0
langsmith>=0.7.0

# ── Document Parsing ──
unstructured[pdf]>=0.12.0
pytesseract>=0.3.10
camelot-py>=1.0.0
PyMuPDF>=1.24.0
pandas>=2.2.0
openpyxl>=3.1.2

# ── ML Models ──
torch-geometric>=2.4.0
scikit-learn>=1.4.0
transformers>=4.38.0
sentence-transformers>=2.7.0

# ── NLP ──
spacy>=3.7.4
gliner>=0.2.0

# ── Research ──
tavily-python>=0.3.0
exa-py>=1.0.0
google-search-results>=2.4.2
beautifulsoup4>=4.12.3
selenium>=4.18.1

# ── Output Generation ──
python-docx>=1.1.0
reportlab>=4.1.0
Jinja2>=3.1.3

# ── Graph Intelligence ──
graphrag>=0.3.0
networkx>=3.2.1

# ── Utilities ──
python-dotenv>=1.2.0
httpx>=0.28.0
aiofiles>=25.1.0
aiosqlite>=0.22.0
